Inspect the file path if no provider is found from the primary title
If no provider site is found by inspecting the Plex-provided media.name, check to see if we have access to the original filepath.

If we do, decode it then grab the parent directory and perform the same trash stripping on it as we do for media.name, then see if we can get a provider match from there.

If that succeeds, there's a chance we might have even more metadata available, so check to see if the parent directory contanined just the site name, and if so, do a further check to see if we can grab more data.