Pass media to search()
Some sites with very picky matching will benefit from closer inspection of the actual filename rather than being forced to work with the media.name that Plex gives us.

Using the file name (and indeed the entire file path) allows inspections like studio/actor.name-scene.name.mp4 where the position of the dash is crucial for the ability to differenciate the actor name from the scene name. We also gain the ability to inspect parent directories for matching clues, if necessary.

No existing providers use this behaviour, but the data is there for the picking, it just needs to be exposed to provider search and we can get started.