Refactor `search` argument to be an instance of class `SearchData`
The main change here is that the former `search` dictionary is deprecated in favour of a `searchData` object, which is an instance of class `SearchData`.

The `SearchData` class lists its properties, and also allows extension with useful methods, such as dateFormat() and durationFormat(), which are already implemented.

All providers are updated to make use of the new object, and instances of `parse(someDateVariable).strftime('%Y-%m-%d')` are updated to make use of the `dateFormat()` method.

A new version of `siteAtkGirlfriends.py` is incoming that makes use of the `durationFormat()` method to allow hugely improved matching.

Secondly, some of the variable names within PhoenixAdult.search() are updated to make them a little more descriptive, especially in the section that deals with generating search strings from the parent directory of the media being matched.

Logging within the search function is tidied up to make it slightly more descriptive and to remove redundancy.

Additionally, instances of `splited` (sic) are replaced with `parts`.