Fix Date Scraping for different conditions
Original date scraping only worked for scenes that were Month Day, Year such as January 01, 2010.  Many scenes are missing the "day" as the movie they are a part of only have a Month, Year such as January, 2010.

This adjustment accounts for either option, but just makes a month, year the first of the month.  So July, 2016 becomes 2016-07-01 in Plex.