Improve plural variant handling
Within the search function, the sceneURL is compared to directURL.
If they are the same, then we replace plurals. ie. sisters become
sister-s. This is problematic because we always turn sisters into
sister-s even if the url is actually sisters. This makes the scene
unsearchable.

Mitigate this case by searching both the plurals substituted and
original sceneURL.

- Make a copy of searchResults called pluralResults
- Process the plurals, at the end append to pluralReults
- Remove duplicates from pluralResults and assign back to searchResults